backend/requirements.txt
fastapi
pydantic
langchain
transformers==4.34.0
torch==2.1.0
uvicorn
langchain_ollama
python-dotenv
